requirements.txt
Flask==3.0.3
psycopg2==2.9.9
requests==2.32.2
flask_sqlalchemy==3.1.1
click==8.1.7
pytest==7.4.4
pytest-flask==1.3.0
